commands/commands/requirements.txt
opencv-python
pycocotools
matplotlib
onnxruntime
onnx
torch
torchvision
tifffile
patchify
scipy
transformers
monai
dataset